Fix loop counter types from int to size_t for 64-bit
Changed loop counters in CVCardAddress to use size_t instead of int when
comparing against .size() or enum constants. This prevents potential
integer overflow issues on 64-bit systems where size_t is 64-bit.

+++ Source/CVCardAddress.cpp
@@ -69,12 +69,12 @@ void CVCardAddress::SetFullAddress(const cdstring& data)
 	}
 	else if (items.size() == eMaxIndex - 1)
 	{
-		for(int i = eExtended; i < eMaxIndex; i++)
+		for(size_t i = eExtended; i < eMaxIndex; i++)
 			mItems[i] = items[i-1];
 	}
 	else
 	{
-		for(int i = 0; i < items.size(); i++)
+		for(size_t i = 0; i < items.size(); i++)
 			mItems[eStreet + i] = items[i];
 	}
 }
